+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/TerrainSlowdown.java
@@ -13,20 +13,23 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.util.Mth;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-@Label(name = "Terrain Slowdown", description = "Slowdown based off the terrain you're walking on. Custom Terrain Slowdown are controlled via json in this feature's folder")
+@Label(name = "Terrain Slowdown", description = "Slowdown based off the terrain entities walking on. Custom Terrain Slowdown are controlled via json in this feature's folder")
 @LoadFeature(module = Modules.Ids.MOVEMENT)
 public class TerrainSlowdown extends JsonFeature {
 	private static final UUID TERRAIN_SLOWDOWN = UUID.fromString("a849043f-b280-4789-bafd-5da8e8e1078e");
@@ -65,15 +68,30 @@ public class TerrainSlowdown extends JsonFeature {
 				|| !event.player.onGround())
 			return;
 
+		applyTerrainSlowdown(event.player, 1f);
+	}
+
+	@SubscribeEvent
+	public void onLivingTick(LivingEvent.LivingTickEvent event) {
+		if (!this.isEnabled()
+				|| !event.getEntity().onGround()
+				|| event.getEntity() instanceof Player
+				|| (event.getEntity().tickCount + event.getEntity().getId()) % 5 != 0)
+			return;
+
+		applyTerrainSlowdown(event.getEntity(), 0.65f);
+	}
+
+	public static void applyTerrainSlowdown(LivingEntity entity, float multiplier) {
 		double onTerrainSlowdown = 0d;
 		int blocks = 0;
-		AABB bb = event.player.getBoundingBox();
+		AABB bb = entity.getBoundingBox();
 		int mX = Mth.floor(bb.minX);
 		int mY = Mth.floor(bb.minY);
 		int mZ = Mth.floor(bb.minZ);
 		for (int x2 = mX; x2 < bb.maxX; x2++) {
 			for (int z2 = mZ; z2 < bb.maxZ; z2++) {
-				BlockState state = event.player.level().getBlockState(BlockPos.containing(x2, event.player.position().y - 0.02d, z2));
+				BlockState state = entity.level().getBlockState(BlockPos.containing(x2, entity.position().y - 0.02d, z2));
 				if (state.isAir())
 					continue;
 				double blockSlowdown = 0d;
@@ -81,7 +99,7 @@ public class TerrainSlowdown extends JsonFeature {
 					if (idTagValue.id.matchesBlock(state.getBlock())) {
 						blockSlowdown = idTagValue.value;
 						if (state.is(BlockTags.ICE) && frostWalkerOnIce) {
-							int lvl = EnchantmentHelper.getEnchantmentLevel(Enchantments.FROST_WALKER, event.player);
+							int lvl = EnchantmentHelper.getEnchantmentLevel(Enchantments.FROST_WALKER, entity);
 							blockSlowdown -= blockSlowdown * (lvl * 0.25f);
 						}
 						blocks++;
@@ -99,7 +117,7 @@ public class TerrainSlowdown extends JsonFeature {
 		for (int x2 = mX; x2 < bb.maxX; x2++) {
 			for (int y2 = mY; y2 < bb.maxY; y2++) {
 				for (int z2 = mZ; z2 < bb.maxZ; z2++) {
-					BlockState state = event.player.level().getBlockState(new BlockPos(x2, y2, z2));
+					BlockState state = entity.level().getBlockState(new BlockPos(x2, y2, z2));
 					if (state.isAir())
 						continue;
 					double blockSlowdown = 0d;
@@ -117,22 +135,20 @@ public class TerrainSlowdown extends JsonFeature {
 		if (blocks != 0)
 			inTerrainSlowdown /= blocks;
 
-		double slowdown = onTerrainSlowdown + inTerrainSlowdown;
+		double slowdown = (onTerrainSlowdown + inTerrainSlowdown) * multiplier;
 
-		AttributeModifier modifier = event.player.getAttribute(Attributes.MOVEMENT_SPEED).getModifier(TERRAIN_SLOWDOWN);
+		AttributeModifier modifier = entity.getAttribute(Attributes.MOVEMENT_SPEED).getModifier(TERRAIN_SLOWDOWN);
 		if (slowdown != 0d) {
 			if (modifier == null) {
-				MCUtils.applyModifier(event.player, Attributes.MOVEMENT_SPEED, TERRAIN_SLOWDOWN, "terrain slowdown", -slowdown, AttributeModifier.Operation.MULTIPLY_BASE, false);
+				MCUtils.applyModifier(entity, Attributes.MOVEMENT_SPEED, TERRAIN_SLOWDOWN, "terrain slowdown", -slowdown, AttributeModifier.Operation.MULTIPLY_BASE, false);
 			}
 			else if (modifier.getAmount() != -slowdown) {
-				event.player.getAttribute(Attributes.MOVEMENT_SPEED).removeModifier(TERRAIN_SLOWDOWN);
-				MCUtils.applyModifier(event.player, Attributes.MOVEMENT_SPEED, TERRAIN_SLOWDOWN, "terrain slowdown", -slowdown, AttributeModifier.Operation.MULTIPLY_BASE, false);
+				entity.getAttribute(Attributes.MOVEMENT_SPEED).removeModifier(TERRAIN_SLOWDOWN);
+				MCUtils.applyModifier(entity, Attributes.MOVEMENT_SPEED, TERRAIN_SLOWDOWN, "terrain slowdown", -slowdown, AttributeModifier.Operation.MULTIPLY_BASE, false);
 			}
 		}
-		else {
-			if (modifier != null) {
-				event.player.getAttribute(Attributes.MOVEMENT_SPEED).removeModifier(TERRAIN_SLOWDOWN);
-			}
-		}
+		else if (modifier != null) {
+            entity.getAttribute(Attributes.MOVEMENT_SPEED).removeModifier(TERRAIN_SLOWDOWN);
+        }
 	}
 }